Move Chrome tests and test server setup to separate packages
