+++ integrationtests/chrome/chrome.go
@@ -1,0 +1,1 @@
+package chrome
